fix(abode): move code to separate functions and clean up code
First step to clean up the codebase and make it more readable:
* Moving main functions, helpers and constants to separate files
* More consistent function api around setters for element lists
* Use proper `Map` instead of using `Object` to create maps
